Created NaturalSQLScriptOrder function
Just a stub so far. But the goal of the branch is:  We want files in a
directory called Table to be run before a directory called View
This function returns higher integer value to files in directory tree
depending on names of parent
directories. The template used is how Redgates scripted database
directory structure looks.